+++ Runtime/Core/GameManagerBase.cs
@@ -1,3 +1,4 @@
+using System;
 using UnityEngine;
 using Wonjeong.Data;
 using Wonjeong.Utils;
@@ -10,8 +11,10 @@ namespace Wonjeong.Core
 
         [SerializeField] private Reporter.Reporter reporter;
         [SerializeField] private GameObject systemCanvas;
-
+        [SerializeField] protected GameObject inspectorContainer;
+        
         private Settings settings;
+        public event Action OnInspectorClosed;
 
         private void Awake()
         {
@@ -40,15 +43,15 @@ namespace Wonjeong.Core
         {
             Cursor.visible = false;
     
-            if (reporter)
+            if (reporter) 
             {
                 if (reporter.show) reporter.show = false;
+                else Debug.LogWarning("[GameManager] reporter is null. Debug UI will be disabled.");
             }
-            else
-            {
-                Debug.LogWarning("[GameManager] reporter is null. Debug UI will be disabled.");
-            }
-
+            
+            if (inspectorContainer) inspectorContainer.SetActive(false);
+            else Debug.LogWarning("[GameManager] inspectorContainer is null.");
+            
             LoadSettings();
         }
 
@@ -71,14 +74,18 @@ namespace Wonjeong.Core
                     reporter.showGameManagerControl = !reporter.showGameManagerControl;
                     if (reporter.show) reporter.show = false;
                 }
-                else
-                {
-                    Debug.LogWarning("[GameManager] reporter is null. Cannot toggle control.");
-                }
+                else Debug.LogWarning("[GameManager] reporter is null. Cannot toggle control.");
             }
-            else if (Input.GetKeyDown(KeyCode.M)) 
+            else if (Input.GetKeyDown(KeyCode.M)) Cursor.visible = !Cursor.visible;
+            else if (Input.GetKeyDown(KeyCode.I))
             {
-                Cursor.visible = !Cursor.visible;
+                if (inspectorContainer) 
+                {
+                    bool isActivating = !inspectorContainer.activeSelf;
+                    inspectorContainer.SetActive(isActivating);
+                    if (!isActivating) OnInspectorClosed?.Invoke(); // 상태가 비활성화(닫힘)로 전환되었을 때만 구독된 이벤트 호출
+                }
+                else Debug.LogWarning("[GameManager] inspectorContainer is null. Cannot toggle inspector.");
             }
         }
     }